% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_lavaan.R
\name{predict_lavaan}
\alias{predict_lavaan}
\title{predict_lavaan}
\usage{
predict_lavaan(fit, newdata = NULL)
}
\arguments{
\item{fit}{model from lavaan sem}

\item{newdata}{new data to use for predictions}
}
\value{
predictions from SEM lavaan model
}
\description{
predict_lavaan
}
